Added x86_64 architecture check for EFI build
Added a check in iso.buildscript so that the x86_64-efi targets gets build only
if the build architecture is x86_64. Now that we are building the efi target on
the target architecture, it is not possible (or necessary really) to provide
the x86_64 EFI executable on the i686 target.